:bug: 修复保存到相册无论如何都会触发

+++ lib/screens/app_screen/receive_file.dart
@@ -65,13 +65,15 @@ class _ReceiveFileState extends State<ReceiveFile> {
           lower.endsWith(".png") ||
           lower.endsWith(".jpeg") ||
           lower.endsWith(".bpm")) {
-        await cross.saveImageToGallery(task.filePath);
-        if (deleteAfterSaveToGallery.value) {
-          await File(task.filePath).delete();
-          await receivingTaskSetRemoved(
-            taskId: task.taskId,
-            reason: 1,
-          );
+        if (saveToGallery.value) {
+          await cross.saveImageToGallery(task.filePath);
+          if (deleteAfterSaveToGallery.value) {
+            await File(task.filePath).delete();
+            await receivingTaskSetRemoved(
+              taskId: task.taskId,
+              reason: 1,
+            );
+          }
         }
       }
     }
